refactor kafka code, following the same structure as other modules, use the async-trait crate
