+++ Cargo.toml
@@ -23,6 +23,7 @@ uuid = { version = "1.12", features = ["v4"] }
 flate2 = "1.1.0"
 anyhow = "1.0.96"
 reqwest = { version = "0.12.12", features = ["json"] }
+async-trait = "0.1.87"
 
 [dev-dependencies]
 criterion = "0.5"
