+++ src/filter/ztf.rs
@@ -3,6 +3,9 @@ use crate::worker_util::WorkerCmd;
 use mongodb::bson::{doc, Document};
 use redis::AsyncCommands;
 use std::{collections::HashMap, error::Error, num::NonZero};
+use tokio::sync::mpsc;
+use tokio::sync::mpsc::error::TryRecvError;
+use tracing::{info, warn};
 
 pub struct ZtfFilter {
     pub id: i32,
@@ -10,6 +13,7 @@ pub struct ZtfFilter {
     pub permissions: Vec<i64>,
 }
 
+#[async_trait::async_trait]
 impl Filter for ZtfFilter {
     async fn build(
         filter_id: i32,
@@ -122,17 +126,14 @@ impl Filter for ZtfFilter {
 
 pub struct ZtfFilterWorker {
     id: String,
-    receiver: std::sync::mpsc::Receiver<WorkerCmd>,
+    receiver: mpsc::Receiver<WorkerCmd>,
     filter_collection: mongodb::Collection<mongodb::bson::Document>,
     alert_collection: mongodb::Collection<mongodb::bson::Document>,
 }
 
+#[async_trait::async_trait]
 impl FilterWorker for ZtfFilterWorker {
-    async fn new(
-        id: String,
-        receiver: std::sync::mpsc::Receiver<WorkerCmd>,
-        config_path: &str,
-    ) -> Self {
+    async fn new(id: String, receiver: mpsc::Receiver<WorkerCmd>, config_path: &str) -> Self {
         let config_file = crate::conf::load_config(&config_path).unwrap();
         let db: mongodb::Database = crate::conf::build_db(&config_file).await;
         let alert_collection = db.collection("ZTF_alerts");
@@ -145,7 +146,7 @@ impl FilterWorker for ZtfFilterWorker {
         }
     }
 
-    async fn run(&self) -> Result<(), Box<dyn Error>> {
+    async fn run(&mut self) -> Result<(), Box<dyn Error>> {
         // query the DB to find the ids of all the filters for ZTF that are active
         let filter_ids: Vec<i32> = self
             .filter_collection
@@ -209,14 +210,38 @@ impl FilterWorker for ZtfFilterWorker {
             .get_multiplexed_async_connection()
             .await
             .unwrap();
+
+        let command_interval: i64 = 500;
+        let mut command_check_countdown = command_interval;
+
         loop {
+            if command_check_countdown == 0 {
+                match self.receiver.try_recv() {
+                    Ok(WorkerCmd::TERM) => {
+                        info!("filterworker {} received termination command", self.id);
+                        break;
+                    }
+                    Err(TryRecvError::Disconnected) => {
+                        warn!(
+                            "filter worker {} receiver disconnected, terminating",
+                            self.id
+                        );
+                        break;
+                    }
+                    Err(TryRecvError::Empty) => {
+                        command_check_countdown = command_interval;
+                    }
+                }
+            }
             for queue in &queues {
                 // get candids from redis
                 let candids: Vec<i64> = con
                     .rpop::<&str, Vec<i64>>(queue.as_str(), NonZero::new(1000))
                     .await
                     .unwrap();
-                if candids.len() == 0 {
+
+                let nb_candids = candids.len();
+                if nb_candids == 0 {
                     continue;
                 }
                 // get the filters that should run on this queue
@@ -241,7 +266,10 @@ impl FilterWorker for ZtfFilterWorker {
                     // push them all at once
                     let _: () = con.lpush(queue_name, out_documents).await.unwrap();
                 }
+                command_check_countdown -= nb_candids as i64;
             }
         }
+
+        Ok(())
     }
 }
